fix logic in checkout and add cart by id

+++ action.php
@@ -12,6 +12,8 @@ if (isset($_POST['pid'])) {
     $pqty = $_POST['pqty'];
     $total_price = $pprice * $pqty;
 
+    $sess_user_id = $_SESSION['id_user'];
+
     $stmt = $conn->prepare('SELECT bouquet_id FROM cart WHERE bouquet_id =?');
     if (!$stmt) {
         echo "Prepare failed: (" . $conn->errno . ") " . $conn->error;
@@ -23,8 +25,8 @@ if (isset($_POST['pid'])) {
     $code = $r['bouquet_id'] ?? '';
 
     if (!$code) {
-	    $query = $conn->prepare('INSERT INTO cart (bouquet_qty, total_price, bouquet_id) VALUES (?, ?, ?)');
-		$query->bind_param('ssi',$pqty, $total_price, $pid);
+	    $query = $conn->prepare('INSERT INTO cart (bouquet_qty, total_price, bouquet_id, user_id) VALUES (?, ?, ?, ?)');
+		$query->bind_param('ssii',$pqty, $total_price, $pid, $sess_user_id);
 		$query->execute();
 
         echo '<script>alert("data sudah masuk")</script>';
@@ -97,14 +99,11 @@ if (isset($_POST['qty'])) {
 
 // Checkout and save customer info in the orders table
 if (isset($_POST['action']) && $_POST['action'] == 'order') {
-    // $name = $_POST['name'];
-    // $email = $_POST['email'];
-    // $phone = $_POST['phone'];
-    // $products = $_POST['products'];
+    $id_user = $_POST['session_user_id'];
+    $userUsername = $_POST['session_user_username'];
     $grand_total = $_POST['grand_total'];
     $address = $_POST['address'];
     $pmode = $_POST['pmode'];
-
 
 	$products = [];
 
@@ -121,31 +120,65 @@ if (isset($_POST['action']) && $_POST['action'] == 'order') {
         }
     }
 
-	// var_dump($products);
+    $order_date = date('Y-m-d');
+
+	// var_dump($_POST);
 	// die;
 
     $data = '';
 
-    $stmt = $conn->prepare('INSERT INTO orders (name, email, phone, address, pmode, products, amount_paid) VALUES (?, ?, ?, ?, ?, ?, ?)');
+    // insert into table orders
+    $stmt = $conn->prepare('INSERT INTO orders(address, pmode, amount_paid, id_user, order_date) VALUES (?, ?, ?, ?, ?)');
     if (!$stmt) {
         echo "Prepare failed: (" . $conn->errno . ") " . $conn->error;
     }
-    $stmt->bind_param('sssssss', $name, $email, $phone, $address, $pmode, $products, $grand_total);
+    $stmt->bind_param('sssis', $address, $pmode, $grand_total, $id_user, $order_date);
     $stmt->execute();
+
+    if($stmt->error){
+        echo "Execute failed: (" . $stmt->errno . ") " . $stmt->error;
+    } else {
+        $order_id = $conn->insert_id;
+        echo "New order created successfully with ID: " . $order_id;
+    }
+
+    // inserts into table orderDetail
+    $orderDetailstmt = $conn->prepare('INSERT INTO order_detail(bouquet_id, qty ,order_id) VALUES(?, ?, ?)');
+
+    foreach ($products as $productKey => $product) {
+        $bouquet_id = $product['bouquet_id'];
+        $bouquet_qty = $product['bouquet_qty']; 
+
+        $orderDetailstmt->bind_param("iii", $bouquet_id, $bouquet_qty, $order_id);
+        if(!$orderDetailstmt->execute()){
+            echo "Error: ". $orderDetailStmt->error();
+        }
+    }
+
+    $orderDetailstmt->close();
+
+
+    $stmt = $conn->prepare('SELECT * FROM cart INNER JOIN tb_user ON cart.user_id = tb_user.id_user');
+    $stmt->execute();
+    $resultUser = $stmt->get_result()->fetch_assoc();
     
-    $stmt2 = $conn->prepare('DELETE FROM cart');
+    $stmt2 = $conn->prepare('DELETE FROM cart WHERE user_id = ?');
+    $stmt2->bind_param('i', $id_user);
+    
     if (!$stmt2) {
         echo "Prepare failed: (" . $conn->errno . ") " . $conn->error;
     }
     $stmt2->execute();
 
+
+
     $data .= '<div class="text-center">
         <h1 class="display-4 mt-2 text-danger">Thank You!</h1>
         <h2 class="text-success">Your Order Placed Successfully!</h2>
         <h4 class="bg-danger text-light rounded p-2">Items Purchased : ' . $products . '</h4>
-        <h4>Your Name : ' . $name . '</h4>
-        <h4>Your E-mail : ' . $email . '</h4>
-        <h4>Your Phone : ' . $phone . '</h4>
+        <h4>Your Name : ' . $resultUser['nama_user'] . '</h4>
+        <h4>Your E-mail : ' . $resultUser['email_user'] . '</h4>
+        <h4>Your Phone : ' . $resultUser['notelp_user'] . '</h4>
         <h4>Total Amount Paid : ' . number_format($grand_total) . '</h4>
         <h4>Payment Mode : ' . $pmode . '</h4>
     </div>';
